Fix autocompleting names in various group chats. Fixes #14026.
Reverts functionality for autocompletes to pre-b18, by always autocompleting the alias. However, to correct against libpurple returning a contact list alias for a contact in the alias field, use the PurpleConvChatBuddy's name field.

The name field contains the formatted name as we're expecting it.

Patch (albeit modified to avoid trying to passive-aggressively rename alias to nick in comments ;)) by paradoxmo.