Add "Log only certain accounts", which lets you log only certain accounts, and add "Log secure chats" which (enabled by default) logs secure chats.
This was prompted by (for the first time evar!) using a Mac at work all day, and wanting my work chats logged but not my personal ones. I know, it breaks account/service/whatever agnosticism, but hey, whatevs. Encrypted chats not being logged has been a popular request for a while, and I was in here anyway.

Defaults to not "log only certain accounts", and accounts default to enabled for logging.

Take the opportunity to backport [da70be37dabb] as well, which renamed "Save open chats when quitting" to "Reopen chats from last time on startup".

Refs #8, fixes #94 (woohoo old tickets).